+++ lib/src/zlib/_inflate_buffer_html.dart
@@ -3,3 +3,7 @@ import 'inflate.dart';
 List<int>? inflateBuffer_(List<int> data) {
   return Inflate(data).getBytes();
 }
+
+bool useNativeZLib_() {
+  return false;
+}