+++ lib/src/zlib/_inflate_buffer_stub.dart
@@ -1,3 +1,7 @@
 List<int>? inflateBuffer_(List<int> data) {
   throw UnsupportedError('inflateBuffer requires html or io.');
 }
+
+bool useNativeZLib_() {
+  return false;
+}